+++ allennlp/sanity_checks/task_checklists/__init__.py
@@ -1,0 +1,4 @@
+from allennlp.sanity_checks.task_checklists.task_suite import TaskSuite
+from allennlp.sanity_checks.task_checklists.sentiment_analysis_suite import (
+    SentimentAnalysisVocabularySuite,
+)